FullDataTools/requirements.txt
opencv-python
h5py
numpy
scikit-learn
open-clip-torch
scipy